# tests/snapshots/openapi.json
{
  "components": {
    "schemas": {
...
      "ArchitectureAnswer_ServiceDependenciesData_": {
        "additionalProperties": false,
        "allOf": [
          {
            "if": {
              "properties": {
                "outcome": {
                  "enum": [
                    "ANSWERED",
                    "PARTIAL"
                  ]
                }
              },
              "required": [
                "outcome"
              ]
            },
            "then": {
              "properties": {
                "data": {
                  "not": {
                    "type": "null"
                  }
                }
              }
            }
          },
          {
            "if": {
              "properties": {
                "observation_context": {
                  "type": "null"
                },
                "tool": {
                  "enum": [
                    "get_architecture_drift",
                    "get_service_dependencies"
                  ]
                }
              },
              "required": [
                "observation_context",
                "tool"
              ]
            },
            "then": {
              "properties": {
                "limitations": {
                  "contains": {
                    "properties": {
                      "code": {
                        "const": "OBSERVATION_CONTEXT_REQUIRED"
                      }
                    },
                    "required": [
                      "code"
                    ]
                  }
                }
              },
              "required": [
                "limitations"
              ]
            }
          },
          {
            "if": {
              "properties": {
                "tool": {
                  "const": "get_evidence"
                }
              },
              "required": [
                "tool"
              ]
            },
            "then": {
              "properties": {
                "claims": {
                  "maxItems": 0
                },
                "evidence_refs": {
                  "maxItems": 0
                },
                "observation_context": {
                  "type": "null"
                }
              }
            }
          },
          {
            "properties": {
              "tool": {
                "const": "get_service_dependencies"
              }
            }
          }
        ],
        "properties": {
          "claims": {
            "items": {
              "discriminator": {
                "mapping": {
                  "DEPLOYED_AS": "#/components/schemas/DeploymentClaim",
                  "DIRECT_DEPENDENCY": "#/components/schemas/DependencyClaim"
                },
                "propertyName": "predicate"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/DependencyClaim"
                },
                {
                  "$ref": "#/components/schemas/DeploymentClaim"
                }
              ]
            },
            "title": "Claims",
            "type": "array"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ServiceDependenciesData"
              },
              {
                "type": "null"
              }
            ]
          },
          "evidence_refs": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Refs",
            "type": "array",
            "uniqueItems": true
          },
          "limitations": {
            "items": {
              "$ref": "#/components/schemas/Limitation"
            },
            "title": "Limitations",
            "type": "array"
          },
          "observation_context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ObservationContextRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "outcome": {
            "$ref": "#/components/schemas/Outcome"
          },
          "producer": {
            "$ref": "#/components/schemas/Producer"
          },
          "schema_version": {
            "const": "0.5",
            "title": "Schema Version",
            "type": "string"
          },
          "snapshot": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SnapshotRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "tool": {
            "enum": [
              "get_service_dependencies",
              "get_evidence",
              "get_architecture_drift"
            ],
            "title": "Tool",
            "type": "string"
          }
        },
        "required": [
          "schema_version",
          "producer",
          "tool",
          "outcome",
          "snapshot",
          "observation_context",
          "data",
          "claims",
          "evidence_refs",
          "limitations"
        ],
        "title": "ArchitectureAnswer[ServiceDependenciesData]",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LocalityEvidenceBody": {
        "additionalProperties": false,
        "properties": {
          "object_operation_id": {
            "anyOf": [
              {
                "pattern": "^[^\\s\\x00-\\x1f\\x7f]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Object Operation Id"
          },
          "refs": {
            "items": {
              "pattern": "^[^\\s\\x00-\\x1f\\x7f]+$",
              "type": "string"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "Refs",
            "type": "array",
            "uniqueItems": true
          },
          "snapshot_id": {
            "pattern": "^aip:snapshot:v1:[0-9a-f]{64}$",
            "title": "Snapshot Id",
            "type": "string"
          }
        },
        "required": [
          "snapshot_id",
          "refs"
        ],
        "title": "LocalityEvidenceBody",
        "type": "object"
      },
...
      "LocalityQueryBody": {
        "additionalProperties": false,
        "properties": {
          "caller_localities": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/WorkloadIdentity"
                },
                "maxItems": 50,
                "minItems": 1,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caller Localities",
            "uniqueItems": true
          },
          "compare": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/WorkloadIdentity"
                },
                "maxItems": 2,
                "minItems": 2,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compare",
            "uniqueItems": true
          },
          "cursor": {
            "anyOf": [
              {
                "maxLength": 2048,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9_-]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cursor"
          },
          "dimensions": {
            "items": {
              "pattern": "^[a-z][a-z0-9_.-]*$",
              "type": "string"
            },
            "minItems": 1,
            "title": "Dimensions",
            "type": "array",
            "uniqueItems": true
          },
          "environment": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[^\\s\\x00-\\x1f\\x7f](?:[^\\x00-\\x1f\\x7f]*[^\\s\\x00-\\x1f\\x7f])?$",
            "title": "Environment",
            "type": "string"
          },
          "first_day": {
            "pattern": "^[0-9]{4}-[0-9]{2}-[0-9]{2}(T[0-9]{2}:[0-9]{2}:[0-9]{2}(\\.[0-9]+)?(Z|[+-][0-9]{2}:[0-9]{2}))?$",
            "title": "First Day",
            "type": "string"
          },
          "last_day": {
            "pattern": "^[0-9]{4}-[0-9]{2}-[0-9]{2}(T[0-9]{2}:[0-9]{2}:[0-9]{2}(\\.[0-9]+)?(Z|[+-][0-9]{2}:[0-9]{2}))?$",
            "title": "Last Day",
            "type": "string"
          },
          "object_operation_id": {
            "anyOf": [
              {
                "pattern": "^[^\\s\\x00-\\x1f\\x7f]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Object Operation Id"
          },
          "provider_service_id": {
            "anyOf": [
              {
                "pattern": "^[^\\s\\x00-\\x1f\\x7f]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Service Id"
          },
          "relation_type": {
            "default": "CALLS",
            "pattern": "^[^\\s\\x00-\\x1f\\x7f]+$",
            "title": "Relation Type",
            "type": "string"
          },
          "snapshot_id": {
            "anyOf": [
              {
                "pattern": "^aip:snapshot:v1:[0-9a-f]{64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot Id"
          },
          "source_selector": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourceSelector"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "environment",
          "first_day",
          "last_day"
        ],
        "title": "LocalityQueryBody",
        "type": "object"
      },
...
      "Outcome": {
        "enum": [
          "ANSWERED",
          "PARTIAL",
          "NOT_ANSWERED"
        ],
        "title": "Outcome",
        "type": "string"
      },
...
      "ServiceDependenciesData": {
        "additionalProperties": false,
        "allOf": [
          {
            "properties": {
              "service": {
                "properties": {
                  "type": {
                    "const": "SERVICE"
                  }
                }
              }
            }
          }
        ],
        "description": "v0.5.0 I3 spec \u00a714.2: deployment is a separate sibling projection, never a dependency -\n`deployment_claim_ids`/`deployment_resolutions` are additive fields alongside the original\ndependency-only shape, not a replacement of it.",
        "properties": {
          "dependency_claim_ids": {
            "items": {
              "type": "string"
            },
            "title": "Dependency Claim Ids",
            "type": "array"
          },
          "deployment_claim_ids": {
            "items": {
              "type": "string"
            },
            "title": "Deployment Claim Ids",
            "type": "array"
          },
          "deployment_resolutions": {
            "items": {
              "$ref": "#/components/schemas/DeploymentResolution"
            },
            "title": "Deployment Resolutions",
            "type": "array"
          },
          "service": {
            "$ref": "#/components/schemas/EntityRef"
          }
        },
        "required": [
          "service",
          "dependency_claim_ids",
          "deployment_claim_ids",
          "deployment_resolutions"
        ],
        "title": "ServiceDependenciesData",
        "type": "object"
      },
      "ServiceDeploymentsView": {
        "additionalProperties": false,
        "description": "v0.5.0 I3 spec \u00a715: the REST-only projection `GET /{service_id}/deployments` returns - not a\npublic MCP contract type (there is no matching MCP tool for this view), so it lives here rather\nthan in `contracts.py`. Every field is extracted from one already-computed\n`get_service_dependencies` answer (spec \u00a715: \"MUST NOT invoke deployment reconciliation, Neo4j\nqueries, or qualification logic independently\") - this model has no validators of its own beyond\nshape, since the answer it projects is already fully validated.",
        "properties": {
          "deployment_claims": {
            "items": {
              "$ref": "#/components/schemas/DeploymentClaim"
            },
            "title": "Deployment Claims",
            "type": "array"
          },
          "deployment_resolutions": {
            "items": {
              "$ref": "#/components/schemas/DeploymentResolution"
            },
            "title": "Deployment Resolutions",
            "type": "array"
          },
          "evidence_refs": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Refs",
            "type": "array"
          },
          "limitations": {
            "items": {
              "$ref": "#/components/schemas/Limitation"
            },
            "title": "Limitations",
            "type": "array"
          },
          "observation_context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ObservationContextRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "schema_version": {
            "const": "0.5",
            "title": "Schema Version",
            "type": "string"
          },
          "service": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EntityRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "snapshot": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SnapshotRef"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "schema_version",
          "snapshot",
          "observation_context",
          "service",
          "deployment_claims",
          "deployment_resolutions",
          "evidence_refs",
          "limitations"
        ],
        "title": "ServiceDeploymentsView",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Architecture Intelligence PoC",
    "version": "snapshot"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/services/{service_id}/dependencies": {
      "get": {
        "description": "`GET /api/services/{service_id}/dependencies` (spec \u00a714.5).",
        "operationId": "get_service_dependencies_api_services__service_id__dependencies_get",
        "parameters": [
          {
            "in": "path",
            "name": "service_id",
            "required": true,
            "schema": {
              "title": "Service Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "environment",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Environment"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "in": "query",
            "name": "snapshot_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Snapshot Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArchitectureAnswer_ServiceDependenciesData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Service Dependencies",
        "tags": [
          "architecture-intelligence"
        ]
      }
    },
    "/api/services/{service_id}/dependencies/by-locality": {
      "post": {
        "description": "`POST /api/services/{service_id}/dependencies/by-locality` (D1, `mode: \"query\"`).",
        "operationId": "get_service_dependencies_by_locality_api_services__service_id__dependencies_by_locality_post",
        "parameters": [
          {
            "in": "path",
            "name": "service_id",
            "required": true,
            "schema": {
              "title": "Service Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocalityQueryBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocalityAnswer"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Service Dependencies By Locality",
        "tags": [
          "architecture-intelligence"
        ]
      }
    },
    "/api/services/{service_id}/dependencies/by-locality/evidence": {
      "post": {
        "description": "`POST /api/services/{service_id}/dependencies/by-locality/evidence` (D1, D11,\n`mode: \"evidence\"`).",
        "operationId": "resolve_scoped_locality_evidence_api_services__service_id__dependencies_by_locality_evidence_post",
        "parameters": [
          {
            "in": "path",
            "name": "service_id",
            "required": true,
            "schema": {
              "title": "Service Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocalityEvidenceBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocalityAnswer"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve Scoped Locality Evidence",
        "tags": [
          "architecture-intelligence"
        ]
      }
    },
    "/api/services/{service_id}/deployments": {
      "get": {
        "description": "`GET /api/services/{service_id}/deployments` (spec \u00a715). Unlike `/dependencies`/`/drift`,\n`environment`/`from`/`to` are required (FastAPI's own required `Query(...)` gives 422 for a\nmissing one for free), and an `UNKNOWN_ENTITY` limitation becomes a real 404 - the one spec-frozen\nexception to this file's \"everything else stays 200\" rule (see the module docstring).",
        "operationId": "get_service_deployments_api_services__service_id__deployments_get",
        "parameters": [
          {
            "in": "path",
            "name": "service_id",
            "required": true,
            "schema": {
              "title": "Service Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "environment",
            "required": true,
            "schema": {
              "title": "Environment",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "From",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "To",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceDeploymentsView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Service Deployments",
        "tags": [
          "architecture-intelligence"
        ]
      }
    },
    "/api/services/{service_id}/drift": {
      "get": {
        "description": "`GET /api/services/{service_id}/drift` (spec \u00a714.5).",
        "operationId": "get_architecture_drift_api_services__service_id__drift_get",
        "parameters": [
          {
            "in": "path",
            "name": "service_id",
            "required": true,
            "schema": {
              "title": "Service Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "environment",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Environment"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "in": "query",
            "name": "snapshot_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Snapshot Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArchitectureAnswer_ArchitectureDriftData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Architecture Drift",
        "tags": [
          "architecture-intelligence"
        ]
      }
    },
...
  }
}